send/edit thread message
